Rename project to "View Strategy".
